+++ app/ingestion/prompts/community-report-prompt.txt
