+++ app/integration/index_manager.py
