+++ tests/test_pdf_processing.py
@@ -1,11 +1,11 @@
 import unittest
 from unittest.mock import patch, MagicMock
-from app.pdf_processing import convert_pdf_page_to_png
+from app.ingestion.pdf_processing import convert_pdf_page_to_png
 
 class TestPdfProcessing(unittest.TestCase):
 
-    @patch('app.pdf_processing.convert_from_path')
-    @patch('app.pdf_processing.os.path.exists')
+    @patch('app.ingestion.pdf_processing.convert_from_path')
+    @patch('app.ingestion.pdf_processing.os.path.exists')
     def test_convert_pdf_page_to_png(self, mock_exists, mock_convert):
         mock_exists.return_value = True
         mock_image = MagicMock()
@@ -15,7 +15,7 @@ class TestPdfProcessing(unittest.TestCase):
         self.assertIsNotNone(result)
         mock_image.save.assert_called_once()
 
-    @patch('app.pdf_processing.os.path.exists')
+    @patch('app.ingestion.pdf_processing.os.path.exists')
     def test_convert_pdf_page_to_png_file_not_exists(self, mock_exists):
         mock_exists.return_value = False
