+++ tests/test_research.py
@@ -1,6 +1,6 @@
 import unittest
 from unittest.mock import patch, MagicMock
-from app.research import create_agent, create_reviewer_agent, create_user_proxy, search, generate_final_conclusion, research_with_data
+from app.query.research import create_agent, create_reviewer_agent, create_user_proxy, search, generate_final_conclusion, research_with_data
 
 class TestResearch(unittest.TestCase):
     def test_create_user_proxy(self):
